Fix example usage
