+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "betty",
     "betweeni",
     "betweenx",
+    "birthdate",
     "clonedeep",
     "consts",
     "datetime",
